Move mask into profile-tool
