+++ cloud/filestore/tools/analytics/profile_tool/lib/mask.h
@@ -1,13 +1,11 @@
 #pragma once
 
-#include "options.h"
-
 #include <cloud/filestore/libs/diagnostics/events/profile_events.ev.pb.h>
 
 #include <library/cpp/eventlog/eventlog.h>
 #include <library/cpp/eventlog/iterator.h>
 
-namespace NCloud::NFileStore::NMaskSensitiveData {
+namespace NCloud::NFileStore::NProfileTool {
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -18,16 +16,25 @@ class TMaskSensitiveData
     int EventMessageNumber = 0;
     const NProto::TProfileLogRecord* MessagePtr{};
 
-    const TOptions& Options;
-
     // Some random string but stable in one session
     TString Seed;
 
 public:
-    explicit TMaskSensitiveData(const TOptions& options);
+    enum class EMode
+    {
+        Nodeid,
+        Hash,
+        Empty,
+    };
+
+private:
+    EMode Mode;
+
+public:
+    explicit TMaskSensitiveData(const EMode& mode);
     bool Advance();
     TString Transform(const TString& str, const ui64 nodeId);
     void MaskSensitiveData(const TString& in, const TString& out);
 };
 
-}   // namespace NCloud::NFileStore::NMaskSensitiveData
+}   // namespace NCloud::NFileStore::NProfileTool
