+++ cloud/filestore/tools/testing/ya.make
@@ -3,5 +3,4 @@ RECURSE(
     loadtest
     open_close_bench
     profile_log
-    mask_sensitive_data
 )
